Move release and production distribution tooling
Keep release metadata, release-summary rendering, and production-distribution selection and validation together with their tests and shell entrypoints. Preserve command syntax, machine-readable results, environment contracts, and exit behavior while changing ownership paths.
