+++ tools/production-distribution/validate-production-distribution.sh
